+++ sycl/source/detail/windows_pi.cpp
@@ -11,12 +11,11 @@
 
 #include <cassert>
 #include <string>
-#include <vector>
 #include <windows.h>
 #include <winreg.h>
 
+#include "detail/windows_os_utils.hpp"
 #include "pi_win_proxy_loader.hpp"
-#include <sycl/detail/windows_os_utils.hpp>
 
 namespace sycl {
 inline namespace _V1 {
@@ -34,10 +33,10 @@ void *loadOsLibrary(const std::string &LibraryPath) {
   if (!SetDllDirectoryA("")) {
     assert(false && "Failed to update DLL search path");
   }
-  auto Result =
-      (void *)LoadLibraryExA(LibraryPath.c_str(), NULL, NULL);
+
+  auto Result = (void *)LoadLibraryExA(LibraryPath.c_str(), NULL, NULL);
   (void)SetErrorMode(SavedMode);
-  if (!SetDllDirectory(nullptr)) {
+  if (!SetDllDirectoryA(nullptr)) {
     assert(false && "Failed to restore DLL search path");
   }
 
@@ -69,11 +68,29 @@ void *getOsLibraryFuncAddress(void *Library, const std::string &FunctionName) {
       GetProcAddress((HMODULE)Library, FunctionName.c_str()));
 }
 
+static std::filesystem::path getCurrentDSODirPath() {
+  wchar_t Path[MAX_PATH];
+  auto Handle =
+      getOSModuleHandle(reinterpret_cast<void *>(&getCurrentDSODirPath));
+  DWORD Ret = GetModuleFileName(
+      reinterpret_cast<HMODULE>(ExeModuleHandle == Handle ? 0 : Handle),
+      reinterpret_cast<LPWSTR>(&Path), sizeof(Path));
+  assert(Ret < sizeof(Path) && "Path is longer than PATH_MAX?");
+  assert(Ret > 0 && "GetModuleFileName failed");
+  (void)Ret;
+
+  BOOL RetCode = PathRemoveFileSpec(reinterpret_cast<LPWSTR>(&Path));
+  assert(RetCode && "PathRemoveFileSpec failed");
+  (void)RetCode;
+
+  return std::filesystem::path(Path);
+}
+
 // Load plugins corresponding to provided list of plugin names.
 std::vector<std::tuple<std::string, backend, void *>>
 loadPlugins(const std::vector<std::pair<std::string, backend>> &&PluginNames) {
   std::vector<std::tuple<std::string, backend, void *>> LoadedPlugins;
-  const std::filesystem::path LibSYCLDir = sycl::detail::getCurrentDSODirPath();
+  const std::filesystem::path LibSYCLDir = getCurrentDSODirPath();
 
   for (auto &PluginName : PluginNames) {
     void *Library = getPreloadedPlugin(LibSYCLDir / PluginName.first);
